cleaner config for various frontend bundles
