+++ web/source/index.js
@@ -33,6 +33,16 @@ function out(name = "") {
 	return path.join(__dirname, "../assets/dist/", name);
 }
 
+module.exports = {out};
+
+const splitCSS = require("./lib/split-css.js");
+
+const bundles = {
+	"./frontend/index.js": "frontend.js",
+	"./panels/admin/index.js": "admin-panel.js",
+	"./panels/user/index.js": "user-panel.js",
+};
+
 const postcssPlugins = [
 	"postcss-import",
 	"postcss-strip-inline-comments",
@@ -41,53 +51,6 @@ const postcssPlugins = [
 	"postcss-custom-prop-vars",
 	"postcss-color-mod-function"
 ].map((plugin) => require(plugin)());
-
-const fromRegex = /\/\* from (.+?) \*\//;
-function splitCSS() {
-	let chunks = [];
-	return new Writable({
-		write: function(chunk, encoding, next) {
-			chunks.push(chunk);
-			next();
-		},
-		final: function() {
-			let stream = chunks.join("");
-			let input;
-			let content = [];
-
-			function write() {
-				if (content.length != 0) {
-					if (input == undefined) {
-						throw new Error("Got CSS content without filename, can't output: ", content);
-					} else {
-						console.log("writing to", out(input));
-						fs.writeFileSync(out(input), content.join("\n"));
-					}
-					content = [];
-				}
-			}
-
-			stream.split("\n").forEach((line) => {
-				if (line.startsWith("/* from")) {
-					let found = fromRegex.exec(line);
-					if (found != null) {
-						write();
-
-						let parts = path.parse(found[1]);
-						if (parts.dir == "css") {
-							input = parts.base;
-						} else {
-							input = found[1].replace(/\//g, "-");
-						}
-					}
-				} else {
-					content.push(line);
-				}
-			});
-			write();
-		}
-	});
-}
 
 const browserifyConfig = {
 	transform: babelify.configure({ presets: [require.resolve("@babel/preset-env"), require.resolve("@babel/preset-react")] }),
@@ -98,14 +61,15 @@ const browserifyConfig = {
 			mode: 'global'
 		}],
 		[require("css-extract"), { out: splitCSS }],
-		[require("factor-bundle"), { outputs: [out("/admin-panel.js"), out("user-panel.js")] }]
+		[require("factor-bundle"), {
+			outputs: Object.values(bundles).map((file) => {
+				return out(file);
+			})
+		}]
 	]
 };
 
-const entryFiles = [
-	'./panels/admin/index.js',
-	'./panels/user/index.js',
-];
+const entryFiles = Object.keys(bundles);
 
 fs.readdirSync(path.join(__dirname, "./css")).forEach((file) => {
 	entryFiles.push(path.join(__dirname, "./css", file));
@@ -125,7 +89,6 @@ const server = budoExpress({
 
 if (server instanceof EventEmitter) {
 	server.on("update", (contents) => {
-	console.log("writing bundle.js to dist/");
-	fs.writeFileSync(out("bundle.js"), contents);
-});
+		fs.writeFileSync(out("bundle.js"), contents);
+	});
 }
